Add nvm to make targets and make recommended
Add nvm to the makefile targets to ensure that the correct node version
is always being used. This is especially helpful to speed up switching
between apps with differing node versions as it avoids the need to run
nvm use.

The readme has also been updated inline with the more prominent use of
nvm to make it clear that nvm is the standard approach to working with
this repo and to better highlight the make targets to use.